+++ src/WavFileReader.cpp
@@ -39,7 +39,7 @@ void WavFileReader::close() {
     }
 }
 
-WavFileReader::Status WavFileReader::read(size_t samplesNum, float * buffer, size_t * samplesRead) {
+WavFileReader::Status WavFileReader::read(size_t samplesNum, double * buffer, size_t * samplesRead) {
     size_t samplesLeft = getSamplesLeft();
     size_t samplesToRead = samplesLeft > samplesNum ? samplesNum : samplesLeft;
     uint8_t * ioBuffer = new uint8_t[(header.bitsPerSample / 8) * samplesNum];
@@ -60,26 +60,26 @@ WavFileReader::Status WavFileReader::read(size_t samplesNum, float * buffer, siz
             if (header.bitsPerSample == 8) {
                 for (size_t i = 0; i < *samplesRead; i++) {
                     sample = ioBuffer[i] - 128;
-                    buffer[i] = (float) sample / (float) INT8_MAX;
+                    buffer[i] = (double) sample / (double) INT8_MAX;
                 }
             } else if (header.bitsPerSample == 16) {
                 for (size_t i = 0; i < *samplesRead; i++) {
                     sample = read16(ioBuffer, i * 2);
                     if (sample & 0x8000)
                         sample = sample | 0xFFFF0000;
-                    buffer[i] = (float) sample / (float) INT16_MAX;
+                    buffer[i] = (double) sample / (double) INT16_MAX;
                 }
             } else if (header.bitsPerSample == 24) {
                 for (size_t i = 0; i < *samplesRead; i++) {
                     sample = read24(ioBuffer, i * 3);
                     if (sample & 0x800000)
                         sample = sample | 0xFF000000;
-                    buffer[i] = (float) sample / (float) INT24_MAX;
+                    buffer[i] = (double) sample / (double) INT24_MAX;
                 }
             } else if (header.bitsPerSample == 32) {
                 for (size_t i = 0; i < *samplesRead; i++) {
                     sample = read32(ioBuffer, i * 4);
-                    buffer[i] = (float) sample / (float) INT32_MAX;
+                    buffer[i] = (double) sample / (double) INT32_MAX;
                 }
             }
         } else if (header.audioFormat == WAV_FILE_AUDIO_FORMAT_IEEE_FLOAT) {
